+++ Source/ImoutoRebirth.Lilin/ImoutoRebirth.Lilin.Infrastructure/Mappers/TagMapper.cs
@@ -16,6 +16,7 @@ public static class TagMapper
             entity.Name,
             entity.HasValue,
             entity.SynonymsArray,
+            entity.Options,
             entity.Count);
     }
 }