+++ Source/ImoutoRebirth.Navigator/ImoutoRebirth.Navigator/Services/Tags/ITagService.cs
@@ -6,7 +6,7 @@ internal interface ITagService
 {
     Task<IReadOnlyCollection<TagType>> GеtTypes();
 
-    Task CreateTag(Guid typeId, string name, bool hasValue, IReadOnlyCollection<string> synonyms);
+    Task CreateTag(Guid typeId, string name, bool hasValue, IReadOnlyCollection<string> synonyms, bool isCounter);
 
     Task<IReadOnlyCollection<Tag>> SearchTags(string name, int count);